Split specific frame types into their own modules
